+++ R/context.R
@@ -6,14 +6,27 @@
 #' `mutate()`
 #'
 #' * `n()` gives the current group size.
+#' * `cur_data()` gives the current data for the current group (exclusing
+#'   grouping variables)
 #' * `cur_group()` gives the group keys, a tibble with one row and one column
 #'   for each grouping variable.
 #' * `cur_group_id()` gives a unique numeric identifier for the current group.
 #' * `cur_column()` gives the current column (in [across()] only).
 #'
+#' See [group_data()] for equivalent functions that return values for all
+#' groups.
+#'
+#' @section data.table:
+#' If you're familiar with data.table:
+#'
+#' * `cur_data()` <-> `.SD`
+#' * `cur_group_id()` <-> `.GRP`
+#' * `cur_group()` <-> `.BY`
+#' * `cur_group_rows()` <-> `.I`
+#'
 #' @examples
 #' df <- tibble(
-#'   g = rep(letters[1:3], 1:3),
+#'   g = sample(rep(letters[1:3], 1:3)),
 #'   x = runif(6),
 #'   y = runif(6)
 #' )
@@ -22,6 +35,10 @@
 #' gf %>% summarise(n = n())
 #'
 #' gf %>% mutate(id = cur_group_id())
+#' gf %>% summarise(row = cur_group_rows())
+#' gf %>% summarise(data = list(cur_group()))
+#' gf %>% summarise(data = list(cur_data()))
+#'
 #' gf %>% mutate(across(everything(), ~ paste(cur_column(), round(.x, 2))))
 #' @name context
 NULL
@@ -29,47 +46,87 @@ NULL
 #' @rdname context
 #' @export
 n <- function() {
-  context_get("..group_size")
+  length(peek_mask("n()")$current_rows())
+}
+
+#' @rdname context
+#' @export
+cur_data <- function() {
+  mask <- peek_mask("cur_data()")
+  data <- mask$full_data()
+  mask$pick(setdiff(names(data), group_vars(data)))
 }
 
 #' @rdname context
 #' @export
 cur_group <- function() {
-  peek_mask()$current_key()
+  peek_mask("cur_group()")$current_key()
 }
 
 #' @rdname context
 #' @export
 cur_group_id <- function() {
-  context_get("..group_number")
+  peek_mask("cur_group_id()")$get_current_group()
+}
+
+#' @rdname context
+#' @export
+cur_group_rows <- function() {
+  peek_mask("cur_group_rows()")$current_rows()
+}
+
+group_labels_details <- function(keys) {
+  glue_collapse(map2_chr(keys, names(keys), function(x, name) {
+    glue("{name} = {value}", value = format_v(x))
+  }), ", ")
+}
+
+cur_group_label <- function() {
+  mask <- peek_mask("cur_group_label()")
+  data <- mask$full_data()
+  if (is_grouped_df(data)) {
+    paste0("group ", cur_group_id(), ": ", group_labels_details(cur_group()))
+  } else if (inherits(data, "rowwise_df")) {
+    paste0("row ", cur_group_id())
+  } else {
+    ""
+  }
 }
 
 #' @rdname context
 #' @export
 cur_column <- function() {
-  context_env[["..current_column_name"]] %||%
-    abort("cur_column() can only be used inside across()")
+  peek_column()
 }
 
 # context accessors -------------------------------------------------------
 
-set_current_column <- function(name) {
-  context_env[["..current_column_name"]] <- name
-}
-
-poke_current_column <- function(name) {
-  old <- context_env[["..current_column_name"]]
-  set_current_column(name)
+context_env <- new_environment()
+context_poke <- function(name, value) {
+  old <- context_env[[name]]
+  context_env[[name]] <- value
   old
 }
+context_peek <- function(name, fun, location = "dplyr verbs") {
+  context_env[[name]] %||%
+    abort(glue("{fun} must only be used inside {location}"))
+}
+context_local <- function(name, value, frame = caller_env()) {
+  old <- context_poke(name, value)
+  expr <- expr(on.exit(context_poke(!!name, !!old), add = TRUE))
+  eval_bare(expr, frame)
+}
 
-context_env <- new_environment()
+peek_column <- function() {
+  context_peek("column", "cur_column()", "across()")
+}
+local_column <- function(x, frame = caller_env()) {
+  context_local("column", x, frame = frame)
+}
 
-context_get <- function(key) {
-  out <- env_get(context_env, key)
-  if (is.null(out)) {
-    expr <- deparse(sys.call(-1))
-    abort(glue("{expr} should only be called inside a dplyr verb"))
-  }
-  out
+peek_mask <- function(fun = "peek_mask()") {
+  context_peek("mask", fun)
+}
+local_mask <- function(x, frame = caller_env()) {
+  context_local("mask", x, frame = frame)
 }