+++ R/mutate.R
@@ -47,7 +47,7 @@
 #'
 #' @export
 #' @inheritParams arrange
-#' @param ... <[`tidy-eval`][dplyr_tidy_eval]> Name-value pairs.
+#' @param ... <[`data-masking`][dplyr_data_masking]> Name-value pairs.
 #'   The name gives the name of the column in the output.
 #'
 #'   The value can be:
@@ -119,6 +119,11 @@
 #' mtcars %>%
 #'   transmute(displ_l = disp / 61.0237)
 #'
+#' # Experimental .remove = "used" allows you to remove variables
+#' # that are used "up" when computing new variables:
+#' mtcars %>%
+#'   mutate(displ_l = disp / 61.0237, .remove = "used")
+#'
 #' # Grouping ----------------------------------------
 #' # The mutate operation may yield different results on grouped
 #' # tibbles because the expressions are computed within groups.
@@ -138,15 +143,50 @@
 #' # Refer to column names stored as strings with the `.data` pronoun:
 #' vars <- c("mass", "height")
 #' mutate(starwars, prod = .data[[vars[[1]]]] * .data[[vars[[2]]]])
-#' # Learn more in ?dplyr_tidy_eval
+#' # Learn more in ?dplyr_data_masking
 mutate <- function(.data, ...) {
   UseMethod("mutate")
 }
 
+#' @rdname mutate
+#' @param .keep \Sexpr[results=rd]{lifecycle::badge("experimental")}
+#'   This is an experimental argument that allows you to control which columns
+#'   from `.data` are retained in the output:
+#'
+#'   * `"all"`, the default, retains all variables.
+#'   * `"used"` keeps any variables used to make new variables; it's useful
+#'     for checking your work as it displays inputs and outputs side-by-side.
+#'   * `"unused"` keeps only existing variables **not** used to make new
+#'     variables.
+#'   * `"none"`, only keeps grouping keys (like [transmute()]).
 #' @export
-mutate.data.frame <- function(.data, ..., .before = NULL, .after = NULL) {
-  cols <- mutate_cols(.data, ...)
+mutate.data.frame <- function(.data, ...,
+                              .keep = c("all", "used", "unused", "none"),
+                              .before = NULL, .after = NULL) {
+  keep <- arg_match(.keep)
+
+  cols <- mutate_cols(.data, ..., .track_usage = keep %in% c("used", "unused"))
   out <- dplyr_col_modify(.data, cols)
+
+  if (keep == "all") {
+    out
+  } else if (keep == "unused") {
+    used <- names(.data)[attr(cols, "used")]
+    keep <- setdiff(names(out), setdiff(used, names(cols)))
+    out <- out[keep]
+  } else if (keep == "used") {
+    used <- names(.data)[attr(cols, "used")]
+    keep <- union(used, names(cols))
+    out <- out[keep]
+  } else if (keep == "none") {
+    keep <- c(
+      # ensure group vars present
+      setdiff(group_vars(.data), names(cols)),
+      # cols might contain NULLs
+      intersect(names(cols), names(out))
+    )
+    out <- out[keep]
+  }
 
   if (!quo_is_null(enquo(.before)) || !quo_is_null(enquo(.after))) {
     # Only change the order of new columns
@@ -155,7 +195,6 @@ mutate.data.frame <- function(.data, ..., .before = NULL, .after = NULL) {
   } else {
     out
   }
-
 }
 
 #' @rdname mutate
@@ -166,22 +205,12 @@ transmute <- function(.data, ...) {
 
 #' @export
 transmute.data.frame <- function(.data, ...) {
-  cols <- mutate_cols(.data, ...)
-  .data <- dplyr_col_modify(.data, cols)
-
-  out_cols <- c(
-    # ensure group vars present
-    setdiff(group_vars(.data), names(cols)),
-    # cols might contain NULLs
-    intersect(names(cols), names(.data))
-  )
-
-  .data[out_cols]
+  mutate(.data, ..., .keep = "none")
 }
 
 # Helpers -----------------------------------------------------------------
 
-mutate_cols <- function(.data, ...) {
+mutate_cols <- function(.data, ..., .track_usage = FALSE) {
   rows <- group_rows(.data)
   # workaround when there are 0 groups
   if (length(rows) == 0L) {
@@ -190,7 +219,7 @@ mutate_cols <- function(.data, ...) {
   rows_lengths <- .Call(`dplyr_vec_sizes`, rows)
 
   o_rows <- vec_order(vec_c(!!!rows, .ptype = integer()))
-  mask <- DataMask$new(.data, caller_env(), rows)
+  mask <- DataMask$new(.data, caller_env(), rows, track_usage = .track_usage)
 
   dots <- enquos(...)
   dots_names <- names(dots)
@@ -221,7 +250,10 @@ mutate_cols <- function(.data, ...) {
       }
 
       if (needs_recycle) {
-        chunks <- map2(chunks, rows_lengths, function(chunk, n) {
+        chunks <- pmap(list(seq_along(chunks), chunks, rows_lengths), function(i, chunk, n) {
+          # set the group so that stop_mutate_recycle_incompatible_size() correctly
+          # identifies it, otherwise it would always report the last group
+          mask$set_current_group(i)
           vec_recycle(chunk, n)
         })
       }
@@ -261,7 +293,7 @@ mutate_cols <- function(.data, ...) {
       stop_mutate_not_vector(index = i, dots = dots, result = e$result)
     },
     vctrs_error_incompatible_type = function(e) {
-      stop_combine(conditionMessage(e), index = i, dots = dots, fn = "mutate")
+      stop_combine(e, index = i, dots = dots, fn = "mutate")
     },
     simpleError = function(e) {
       stop_eval_tidy(e, index = i, dots = dots, fn = "mutate")
@@ -270,5 +302,6 @@ mutate_cols <- function(.data, ...) {
 
   is_zap <- map_lgl(new_columns, inherits, "rlang_zap")
   new_columns[is_zap] <- rep(list(NULL), sum(is_zap))
+  attr(new_columns, "used") <- mask$get_used()
   new_columns
 }