+++ tests/testthat/test-relocate.R
@@ -19,3 +19,15 @@ test_that("can only supply one of .before and .after", {
   df <- tibble(x = 1)
   expect_error(relocate(df, .before = 1, .after = 1), "only one")
 })
+
+test_that("before and after are defused with context", {
+  local_fn <- identity
+  expect_identical(
+    names(relocate(mtcars, 3, .before = local_fn(5))),
+    names(relocate(mtcars, 3, .before = 5))
+  )
+  expect_identical(
+    names(relocate(mtcars, 3, .after = local_fn(5))),
+    names(relocate(mtcars, 3, .after = 5))
+  )
+})